add some new vis
