+++ D3JS_Test/d3_test_grid_dynamic.html
@@ -1,0 +1,111 @@
+<html>
+
+<head>
+
+	<script src="https://cdnjs.cloudflare.com/ajax/libs/d3/3.5.6/d3.min.js" charset="utf-8"></script>
+	<script src="https://code.jquery.com/jquery-1.11.3.min.js"></script>
+	<script src="https://cdn.firebase.com/js/client/2.3.1/firebase.js"></script>
+
+	<script>
+
+		$().ready(function (){
+
+			var svgSection = d3
+					.select(".datavis")
+					.append("svg")
+					.attr("width", 1200)
+					.attr("height", 900);
+
+			var color = d3.scale.category20();
+
+			var firebaseRef = new Firebase("https://ikdde-team6.firebaseio.com/");
+
+			firebaseRef.child("agriculturePrice").on("value", function(snapshot) {
+
+				var data = snapshot.val();
+				data = data
+					.filter(function (it){return it.pricelist;});
+
+				var timedPadding = {}
+				data[0].pricelist.forEach(function (it){
+					timedPadding[it.item] = {r:[], value:[]};
+				});
+
+				data.forEach(function (iter){
+
+					var list = iter.pricelist.slice(0, 20);
+
+					list.forEach(function (e){
+						timedPadding[e.item].r.push(e.price / 4);
+						timedPadding[e.item].value.push(e.value);
+					});
+
+				});
+
+				var timedNode = [];
+				for(var i in timedPadding){
+					timedNode.push({
+						item: i,
+						r: timedPadding[i].r, 
+						value: timedPadding[i].value
+					});
+				}
+
+				var count = 0
+				var groupSections = svgSection
+					.selectAll("g")
+					.data(timedNode)
+					.enter() 
+					.append("g")
+					.attr("transform", function (d) {
+						var col = count % 20;
+						var row = Math.floor(count / 20);
+						count++;
+						return "translate(" + (col * 50 + 60) + ", " + (row * 50 + col * 20 + 60) + ")";
+					});
+
+				var circleInGroup = groupSections
+					.append("circle")
+					.attr("cx", 0)
+					.attr("cy", 0)
+					.attr("r",  function (d) {
+						return d.r[0];
+					})
+					.style("fill", function(d) {
+						return color(d.r[0]);
+					});
+
+				var textInGroup = groupSections.append("text")
+					.text(function (d) {
+						return d.item;
+					})
+					.attr("x", -20)
+					.attr("y", 5);
+
+				var len = timedNode[0].r.length;
+				var i = 0;
+				setInterval(function () {
+					circleInGroup
+						.transition()
+						.attr("r", function (d) {
+							return d.r[i];
+						})
+						.duration(100)
+					if(i < len-1) i++;
+					console.log(len - i);
+					$("#date").html(i);
+				}, 120);
+				
+			});
+
+		});
+
+	</script>
+</head>
+
+<body>
+date <span id="date"></span><br>
+<div class="datavis"></div>
+</body>
+
+</html>
